test: split ECharts conditional formatting coverage
